+++ CHANGES.rst
@@ -4,6 +4,7 @@ Release history
 2.2.1 (XXXX-XX-XX)
 ++++++++++++++++++
 
+- Re-structure and complete the documentation.
 - Build SAFE format for Sentinel products from AWS STAC API providers, fixes
   `GH-171 <https://github.com/CS-SI/eodag/issues/171>`_
 - Deprecate ``EODataAccessGateway.load_stac_items``, use the ``StaticStacSearch`` search plugin instead